requirements.txt
chainlit
pyjwt[crypto]
dotmap